[수정] 회원가입 페이지 컴포넌트 분리(약관동의, 가입폼)
